+++ modules/gitlab-backend/lib/index.js
@@ -1,22 +1,23 @@
 'use strict';
 
-var Promise = require('bluebird');
+const { GitLabGroupService } = require('gitter-web-gitlab');
 
 function GitLabBackend(user, identity) {
   this.user = user;
   this.identity = identity;
 }
 
-GitLabBackend.prototype.getEmailAddress = Promise.method(function(/*preferStoredEmail*/) {
+GitLabBackend.prototype.getEmailAddress = function(/*preferStoredEmail*/) {
   return this.identity.email;
-});
+};
 
-GitLabBackend.prototype.findOrgs = Promise.method(function() {
-  return [];
-});
+GitLabBackend.prototype.findOrgs = function() {
+  const gitlabGroupService = new GitLabGroupService(this.user);
+  return gitlabGroupService.getGroups();
+};
 
-GitLabBackend.prototype.getProfile = Promise.method(function() {
+GitLabBackend.prototype.getProfile = function() {
   return { provider: 'gitlab' };
-});
+};
 
 module.exports = GitLabBackend;