+++ modules/gitlab/lib/index.js
@@ -2,5 +2,6 @@
 
 module.exports = {
   GitLabIssuableService: require('./issuable-service'),
-  GitLabIssuableStateService: require('./issuable-state-service')
+  GitLabIssuableStateService: require('./issuable-state-service'),
+  GitLabGroupService: require('./group-service')
 };