+++ public/js/views/community-create/step-constants.js
@@ -3,8 +3,6 @@
 var stepConstants = {
   MAIN: 'main',
   GITHUB_PROJECTS: 'github_projects',
-  INVITE: 'invite',
-  INVITE_CONFIRMATION: 'invite_confirmation',
   OVERVIEW: 'overview'
 };