+++ public/js/vue/thread-message-feed/components/chat-item-actions-test.js
@@ -13,9 +13,25 @@ describe('thread-message-feed chat-item-actions', () => {
     context.mockImplementation(() => ({}));
   });
 
-  it('matches snapshot', () => {
-    const { wrapper } = mount(ChatItemActions, { message });
-    expect(wrapper.element).toMatchSnapshot();
+  it('clicking quote option triggers action', () => {
+    const { wrapper, stubbedActions } = mount(ChatItemActions, { message });
+    wrapper.find('.js-chat-item-quote-action').trigger('click');
+    expect(stubbedActions.threadMessageFeed.quoteMessage).toHaveBeenCalled();
+  });
+
+  describe('user is not the author of the message', () => {
+    it('matches snapshot', () => {
+      const { wrapper } = mount(ChatItemActions, { message });
+      expect(wrapper.element).toMatchSnapshot();
+    });
+
+    it('triggers report action if the report option is clicked', () => {
+      const { wrapper, stubbedActions } = mount(ChatItemActions, { message });
+      // removing the original implementation to prevent an API call
+      stubbedActions.threadMessageFeed.reportMessage.mockImplementation(() => {});
+      wrapper.find('.js-chat-item-report-action').trigger('click');
+      expect(stubbedActions.threadMessageFeed.reportMessage).toHaveBeenCalled();
+    });
   });
 
   describe('user can delete the message', () => {
@@ -34,8 +50,30 @@ describe('thread-message-feed chat-item-actions', () => {
       const { wrapper, stubbedActions } = mount(ChatItemActions, { message });
       // removing the original implementation to prevent an API call
       stubbedActions.threadMessageFeed.deleteMessage.mockImplementation(() => {});
-      wrapper.find('.popover-item__action').trigger('click');
+      wrapper.find('.js-chat-item-delete-action').trigger('click');
       expect(stubbedActions.threadMessageFeed.deleteMessage).toHaveBeenCalled();
+    });
+  });
+
+  describe('user can edit the message', () => {
+    const testMessage = createSerializedMessageFixture({ sent: new Date().toISOString() });
+    beforeEach(() => {
+      context.getUserId.mockImplementation(function() {
+        return testMessage.fromUser.id;
+      });
+    });
+
+    it('matches snapshot', () => {
+      const { wrapper } = mount(ChatItemActions, { message: testMessage });
+      expect(wrapper.element).toMatchSnapshot();
+    });
+
+    it('triggers edit action when edit option is clicked', () => {
+      const { wrapper, stubbedActions } = mount(ChatItemActions, { message: testMessage });
+      // removing the original implementation to prevent an API call
+      stubbedActions.threadMessageFeed.editMessage.mockImplementation(() => {});
+      wrapper.find('.js-chat-item-edit-action').trigger('click');
+      expect(stubbedActions.threadMessageFeed.editMessage).toHaveBeenCalled();
     });
   });
 });